requirements_dl.txt
torch
torchvision
facenet-pytorch
numpy>=1.24.0
opencv-python>=4.7.0
Pillow>=9.5.0
face_recognition>=1.3.0
dlib>=19.24.0
cmake>=3.26.0
tk>=0.1.0
